Finish org so its more like wpilib
