+++ src/lemonlib/drive/__init__.py
@@ -1,3 +1,3 @@
-from .vector2d import Vector2d
-from .swagdrive import SwagDrive
-from .killoughdrive import KilloughDrive
+from ._vector2d import Vector2d
+from ._swagdrive import SwagDrive
+from ._killoughdrive import KilloughDrive
