+++ src/lemonlib/simulation/__init__.py
@@ -1,2 +1,3 @@
-from .kilosim import KilloughDriveSim
-from .lemoninputsim import LemonInputSim
+from ._kilosim import KilloughDriveSim
+from ._lemoninputsim import LemonInputSim
+from ._lemoncamsim import LemonCameraSim
