+++ src/lemonlib/smart/__init__.py
@@ -1,3 +1,3 @@
-from .controller import SmartController
-from .preference import SmartPreference
-from .profile import SmartProfile
+from ._controller import SmartController
+from ._preference import SmartPreference
+from ._profile import SmartProfile
